refactor(qos): Move policies into their own packages by type
