+++ orca-qos/src/main/kotlin/com/netflix/spinnaker/orca/qos/promotionpolicy/NaivePromotionPolicy.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.spinnaker.orca.qos.strategy.naive
+package com.netflix.spinnaker.orca.qos.promotionpolicy
 
 import com.netflix.spinnaker.orca.pipeline.model.Execution
 import com.netflix.spinnaker.orca.qos.PromotionPolicy
@@ -22,7 +22,7 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.stereotype.Component
 
 @Component
-@ConditionalOnProperty("qos.strategy.naive.enabled")
+@ConditionalOnProperty("qos.promotionPolicy.naive.enabled")
 class NaivePromotionPolicy : PromotionPolicy {
 
   override fun apply(candidates: List<Execution>) =
